# metal : fuse add, mul