target #14629 

Fuse `GGML_OP_ADD` and `GGML_OP_MUL`

```bash
LLAMA_SET_ROWS=1 ./scripts/compare-commits.sh master gg/metal-fuse-add -m ./models/qwen3-30b-a3b/ggml-model-q8_0.gguf -m models/gemma-3-4b/ggml-model-q8_0.gguf -fa 1 -t 1
```

| Model                 | Test   |   t/s master |   t/s gg/metal-fuse-add |   Speedup |
|:----------------------|:-------|-------------:|------------------------:|----------:|
| gemma3 4B Q8_0        | pp512  |      2444.84 |                 2494.63 |      1.02 |
| gemma3 4B Q8_0        | tg128  |        90.39 |                   96.76 |      1.07 |
| qwen3moe 30B.A3B Q8_0 | pp512  |      1362.92 |                 1420.74 |      1.04 |
| qwen3moe 30B.A3B Q8_0 | tg128  |        70.12 |                   76.68 |      1.09 |

#### Testing

```bash
make -j && GGML_METAL_FUSION_DEBUG=2 ./bin/test-backend-ops -o RMS_NORM_MUL_ADD -b Metal
```

```
Backend 1/3: Metal
  Device description: Apple M4 Max
  Device memory: 28753 MB (28747 MB free)

ggml_metal_encode_node: fuse: RMS_NORM + MUL + ADD
  RMS_NORM_MUL_ADD(type=f32,ne=[64,5,4,3],eps=0.000000): OK
ggml_metal_encode_node: fuse: RMS_NORM + MUL + ADD
  RMS_NORM_MUL_ADD(type=f32,ne=[64,5,4,3],eps=0.000001): OK
ggml_metal_encode_node: fuse: RMS_NORM + MUL + ADD
  RMS_NORM_MUL_ADD(type=f32,ne=[64,5,4,3],eps=0.000100): OK
ggml_metal_encode_node: fuse: RMS_NORM + MUL + ADD
  RMS_NORM_MUL_ADD(type=f32,ne=[64,5,4,3],eps=0.100000): OK
ggml_metal_encode_node: fuse: RMS_NORM + MUL + ADD
  RMS_NORM_MUL_ADD(type=f32,ne=[64,5,4,3],eps=1.000000): OK
  6543/6543 tests passed
  Backend Metal: OK
ggml_backend_metal_device_rel: fused ADD: 5
ggml_backend_metal_device_rel: fused MUL: 5
```

---

- [x] Disable with env variable
- [x] Print fuse stats
- [x] Fuse with norms, cpys, etc.
- [x] Cleaner kernel impl?